* Accept bare chords like `C` to mean a C major chord. * (#5) Always print the help screen if no arguments are provided. * Split the damn thing into a bunch of files so the imports can be cleaned up. * Bump to GHC 9.2.5. * Print 4 chords in a row if vertical, or 6 in a row if horizontal.
