+++ ChangeLog.md
@@ -3,7 +3,12 @@
 # 1.3.0
 
 * Fixed the lack of the B note in the readme
-* Use an ASCII "o" character for the fingerings instead of "●" which breaks some terminals (#3)
+* ([#3](https://github.com/DataKinds/cherchord/issues/3)) Use an ASCII "o" character for the fingerings instead of "●" which breaks some terminals 
+* Split the damn thing into a bunch of files so the imports can be cleaned up.
+* Bump to GHC 9.2.5.
+* Print 4 chords in a row if vertical, or 6 in a row if horizontal.
+* ([#5](https://github.com/DataKinds/cherchord/issues/5)) Always print the help screen if no arguments are provided.
+* Accept bare chords like `C` to mean a C major chord.
 
 # 1.2.0
