+++ app/Main.hs
@@ -1,137 +1,18 @@
-{-# LANGUAGE ApplicativeDo #-}
-
 module Main where
 
-import Chords
-import Text.Megaparsec
-import Text.Megaparsec.Debug
-import Text.Megaparsec.Char
-import Data.Void
 import Data.List
 import Data.List.Split
 import System.IO
-import Options.Applicative
+import Options.Applicative (customExecParser, prefs, showHelpOnEmpty)
 import Data.Semigroup ((<>))
 import System.Console.ANSI
-import Paths_cherchord (version)
-import Data.Version (showVersion)
+import Text.Megaparsec (parse)
+import Text.Megaparsec.Error (errorBundlePretty)
+import qualified Data.Text as Text (unpack)
 
-type InputParser = Parsec Void String
-data AppOptions = AppOptions {
-  chordIn :: String,
-  fingerStretch :: Int,
-  amountToPrint :: Int,
-  isHorizontal :: Bool,
-  instrument :: Fretboard
-}
-
-parseNote :: InputParser Note
-parseNote = foldr1 (<|>) $ try <$> [
-  string "Ab" >> return Ab,
-  string "A" >> return A,
-  string "Bb" >> return Bb,
-  string "B" >> return B,
-  string "C" >> return C,
-  string "Db" >> return Db,
-  string "D" >> return D,
-  string "Eb" >> return Eb,
-  string "E" >> return E,
-  string "F" >> return F,
-  string "Gb" >> return Gb,
-  string "G" >> return G]
-
-parseKey :: InputParser String
-parseKey = foldr1 (<|>) $ try <$> [
-  string "maj",
-  string "min",
-  string "dim",
-  string "aug",
-  string "sus2",
-  string "sus4"]
-
-parseBaseChord :: InputParser Chord
-parseBaseChord = do
-  note <- parseNote
-  key <- parseKey
-  case key of
-    "maj" -> return $ maj note
-    "min" -> return $ Chords.min note
-    "dim" -> return $ dim note
-    "aug" -> return $ aug note
-    "sus2" -> return $ sus2 note
-    "sus4" -> return $ sus4 note
-
-parseModifiedChord :: InputParser Chord
-parseModifiedChord = do
-  base <- parseBaseChord
-  modifiers <- Text.Megaparsec.many (try parseSlash <|> try parseAdd)
-  return $ foldr (.) id modifiers base
-  where
-    parseSlash :: InputParser (Chord -> Chord)
-    parseSlash = do
-      string "/"
-      note <- parseNote
-      return $ slash note
-    parseAdd :: InputParser (Chord -> Chord)
-    parseAdd = do
-      string "add"
-      num <- Text.Megaparsec.some digitChar
-      return $ add (read num)
-
-validInstruments :: String -> Either String Fretboard
-validInstruments "guitar" = Right guitar
-validInstruments "ukulele" = Right ukulele
-validInstruments "mandolin" = Right mandolin
-validInstruments "bouzouki" = Right bouzouki
-validInstruments "baglamas" = Right baglamas
-validInstruments i =
-  case parse parseInstrument "instrument" i of
-    Left bundle -> Left "Valid instruments are: guitar, ukulele, mandolin, bouzouki, baglamas"
-    Right fretboard -> Right fretboard
-  where
-    parseInstrument :: InputParser Fretboard
-    parseInstrument = do
-      let parseFret = do
-            note <- parseNote
-            num <- Text.Megaparsec.some digitChar
-            optional $ char ','
-            return $ Fret note (read num)
-      Text.Megaparsec.some parseFret
-      
-parseOptions :: Parser AppOptions
-parseOptions = AppOptions <$>
-  strArgument (metavar "CHORD") <*>
-  Options.Applicative.option auto (
-    long "finger-stretch" <>
-    short 'f' <>
-    value 3 <>
-    showDefault <>
-    metavar "FRETS" <>
-    help "How far can your fingers stretch?") <*>
-  Options.Applicative.option auto (
-    long "print-n" <>
-    short 'p' <>
-    value 10000 <>
-    showDefault <>
-    metavar "FINGERINGS" <>
-    help "How many fingerings to print?") <*>
-  Options.Applicative.switch (
-    long "horizontal" <>
-    help "Should we print the chords horizontally? By default, they are printed vertically.") <*>
-  Options.Applicative.option (eitherReader validInstruments) (
-    long "instrument" <>
-    short 'i' <>
-    value guitar <>
-    showDefault <>
-    metavar "INSTRUMENT" <>
-    help "What instrument to show chord diagrams for? Valid instruments are: guitar, ukulele, mandolin, bouzouki, baglamas or a comma-delimited list of notes followed by numbers.\nExample: a guitar can be defined as E16,A16,D16,G16,B16,E16.")
-
-parserInfoOptions :: ParserInfo AppOptions
-parserInfoOptions = info (parseOptions <**> helper) (
-  fullDesc <>
-  progDesc "Searches for chord fingerings on a given instrument." <>
-  header "cherchord -- find your fingers" <>
-  footer ("cherchord v" ++ showVersion version ++ " (c) 2019 https://github.com/aearnus/cherchord"))
+import qualified Chords as C
+import qualified ArgumentParser as A
+import qualified ChordParser as P
 
 horizConcat :: [String] -> String
 horizConcat = foldr1 horizConcatOne
@@ -155,11 +36,14 @@ horizConcat = foldr1 horizConcatOne
 
 main :: IO ()
 main = do
-  opts <- execParser parserInfoOptions
-  case parse parseModifiedChord "chord" (chordIn opts) of
+  opts <- customExecParser (prefs showHelpOnEmpty) A.parserInfoOptions
+  case parse P.parseModifiedChord "chord" (A.chordIn opts) of
     Left bundle -> putStrLn (errorBundlePretty bundle)
     Right chord -> do
-      let chords = search chord (fingerStretch opts) (instrument opts)
+      let chords = C.search chord (A.fingerStretch opts) (A.instrument opts)
+          printCount = A.amountToPrint opts
+          whichShow = if A.isHorizontal opts then C.showHorizontally else show
+          horizontalCount = if A.isHorizontal opts then 6 else 4
       setSGR [SetColor Foreground Dull Blue]
       putStr "found "
       setSGR [SetColor Foreground Vivid Blue]
@@ -167,10 +51,9 @@ main = do
       setSGR [SetColor Foreground Dull Blue]
       putStr " unique fingerings for the chord "
       setSGR [SetColor Foreground Vivid Blue]
-      putStr (chordIn opts)
-      putStrLn . (\s -> " (" ++ s ++ ")") . show $ upperNotes chord
+      putStr (A.chordIn opts)
+      putStrLn . (\s -> " (" ++ s ++ ")") . show $ C.upperNotes chord
       setSGR [SetColor Foreground Dull Blue]
-      putStrLn $ "printing out " ++ show (Prelude.min (amountToPrint opts) (length chords)) ++ " of them...\n"
+      putStrLn $ "printing out " ++ show (min printCount (length chords)) ++ " of them...\n"
       setSGR [Reset]
-      let whichShow = if isHorizontal opts then showHorizontally else show
-      putStrLn . intercalate "\n" . map horizConcat . chunksOf 3 . map whichShow . take (amountToPrint opts) $ chords
+      putStrLn . intercalate "\n" . map horizConcat . chunksOf horizontalCount . map whichShow . take printCount $ chords
