+++ package.yaml
@@ -1,6 +1,6 @@
 name:                cherchord
-version:             1.2.0
-github:              "aearnus/cherchord"
+version:             1.3.0
+github:              "DataKinds/cherchord"
 license:             BSD3
 author:              "Tyler"
 maintainer:          "tslimkemann42@gmail.com"
@@ -17,7 +17,7 @@ extra-source-files:
 # To avoid duplicated efforts in documentation and dealing with the
 # complications of embedding Haddock markup inside cabal files, it is
 # common to point users to the README.md file.
-description:         Please see the README on GitHub at <https://github.com/aearnus/cherchord>
+description:         Please see the README on GitHub at <https://github.com/DataKinds/cherchord>
 
 dependencies:
 - base >= 4.7 && < 5
@@ -26,6 +26,8 @@ dependencies:
 - optparse-applicative
 - split
 - ansi-terminal
+- neat-interpolation
+- text
 
 library:
   source-dirs: src
@@ -38,6 +40,9 @@ executables:
     - -threaded
     - -rtsopts
     - -with-rtsopts=-N
+    default-extensions:
+      - ApplicativeDo
+      - QuasiQuotes
     dependencies:
     - cherchord
