+++ stack.yaml
@@ -18,7 +18,7 @@
 #
 # resolver: ./custom-snapshot.yaml
 # resolver: https://example.com/snapshots/2018-01-01.yaml
-resolver: lts-13.15
+resolver: lts-20.8
 
 # User packages to be built.
 # Various formats can be used as shown in the example below.
